Move special pages to folder and change hook handler
