+++ includes/MirahezeIRCRCFeedFormatter.php
@@ -1,7 +1,5 @@
 <?php
 
-namespace Miraheze\MirahezeMagic;
-
 /**
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
